requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pandas
psycopg2
